+++ build_runner/lib/src/build_plan/build_packages.dart
@@ -33,8 +33,11 @@ class BuildPackages implements AssetPathProvider {
   /// cycles.
   final BuiltList<String> orderedPackages;
 
-  /// When building a single package: the package `build_runner` was launched
-  /// in, which is also the current directory package and the [outputRoot].
+  /// The package `build_runner` was launched in, which is also the current
+  /// directory package.
+  final String currentPackage;
+
+  /// When building a single package, [currentPackage].
   ///
   /// When building a workspace, `null`.
   final String? singleOutputPackage;
@@ -71,6 +74,7 @@ class BuildPackages implements AssetPathProvider {
   final BuiltSetMultimap<String, String> _peerPackages;
 
   BuildPackages({
+    required this.currentPackage,
     required this.singleOutputPackage,
     required this.outputRoot,
     required this.packages,
@@ -83,6 +87,7 @@ class BuildPackages implements AssetPathProvider {
        _peerPackages = buildPackages;
 
   factory BuildPackages.compute({
+    required String currentPackage,
     String? singlePackageToBuild,
     required String outputRoot,
     required BuiltMap<String, BuildPackage> packages,
@@ -124,6 +129,7 @@ class BuildPackages implements AssetPathProvider {
     final buildPackages = _computePeers(outputPackages, transitiveDependencies);
 
     return BuildPackages(
+      currentPackage: currentPackage,
       singleOutputPackage: singlePackageToBuild,
       outputRoot: outputRoot,
       packages: packages,
@@ -142,6 +148,7 @@ class BuildPackages implements AssetPathProvider {
     String package,
     Iterable<BuildPackage> packages,
   ) => BuildPackages.compute(
+    currentPackage: package,
     singlePackageToBuild: package,
     outputRoot: package,
     packages: {for (final package in packages) package.name: package}.build(),
@@ -150,10 +157,12 @@ class BuildPackages implements AssetPathProvider {
   /// Creates [BuildPackages] from [BuildPackage]s for a workspace build with
   /// workspace name [workspace].
   @visibleForTesting
-  factory BuildPackages.workspaceBuild(
-    String workspace,
-    Iterable<BuildPackage> packages,
-  ) => BuildPackages.compute(
+  factory BuildPackages.workspaceBuild({
+    required String currentPackage,
+    required String workspace,
+    required Iterable<BuildPackage> packages,
+  }) => BuildPackages.compute(
+    currentPackage: currentPackage,
     outputRoot: workspace,
     packages: {for (final package in packages) package.name: package}.build(),
   );