+++ build_runner/test/integration_tests/build_command_define_test.dart
@@ -69,11 +69,14 @@ class TestBuilder implements Builder {
     );
 
     // Default dev build.
-    await tester.run('root_pkg', 'dart run build_runner build');
+    await tester.run('root_pkg', 'dart run build_runner build --force-jit');
     expect(tester.read('root_pkg/web/a.txt.copy'), 'a(default dev)');
 
     // Default release build.
-    await tester.run('root_pkg', 'dart run build_runner build --release');
+    await tester.run(
+      'root_pkg',
+      'dart run build_runner build --force-jit --release',
+    );
     expect(tester.read('root_pkg/web/a.txt.copy'), 'a(default release)');
 
     // Configure via `build.yaml`.
@@ -89,16 +92,19 @@ targets:
         release_options:
           extra_content: "(yaml release)"
 ''');
-    await tester.run('root_pkg', 'dart run build_runner build');
+    await tester.run('root_pkg', 'dart run build_runner build --force-jit');
     expect(tester.read('root_pkg/web/a.txt.copy'), 'b(yaml dev)');
 
-    await tester.run('root_pkg', 'dart run build_runner build --release');
+    await tester.run(
+      'root_pkg',
+      'dart run build_runner build --force-jit --release',
+    );
     expect(tester.read('root_pkg/web/a.txt.copy'), 'b(yaml release)');
 
     // Override with `--define`.
     await tester.run(
       'root_pkg',
-      'dart run build_runner build '
+      'dart run build_runner build --force-jit '
           '--define=builder_pkg:test_builder=copy_from=root_pkg|web/a.txt',
     );
     expect(tester.read('root_pkg/web/a.txt.copy'), 'a(yaml dev)');
@@ -106,7 +112,7 @@ targets:
     // Override with `--define` and `--release`.
     await tester.run(
       'root_pkg',
-      'dart run build_runner build --release '
+      'dart run build_runner build --force-jit --release '
           '--define=builder_pkg:test_builder=copy_from=root_pkg|web/a.txt',
     );
     expect(tester.read('root_pkg/web/a.txt.copy'), 'a(yaml release)');
@@ -129,7 +135,7 @@ global_options:
     options:
       extra_content: "(global)"
 ''');
-    await tester.run('root_pkg', 'dart run build_runner build');
+    await tester.run('root_pkg', 'dart run build_runner build --force-jit');
     expect(tester.read('root_pkg/web/a.txt.copy'), 'b(global)');
 
     // Change to a workspace.
@@ -144,7 +150,7 @@ global_options:
 
     // Build with --workspace. Package options from `root_pkg/build.yaml`
     // still apply, but global options from `root_pkg/build.yaml` don't.
-    await tester.run('', 'dart run build_runner build --workspace');
+    await tester.run('', 'dart run build_runner build --force-jit --workspace');
     expect(tester.read('root_pkg/web/a.txt.copy'), 'b(yaml dev)');
 
     // Global options from workspace root `build.yaml` are used.
@@ -154,7 +160,7 @@ global_options:
     options:
       extra_content: "(workspace global)"
 ''');
-    await tester.run('', 'dart run build_runner build --workspace');
+    await tester.run('', 'dart run build_runner build --force-jit --workspace');
     expect(tester.read('root_pkg/web/a.txt.copy'), 'b(workspace global)');
   });
 }